// src/model/generated/vehiclePublic.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * PASS2RENT widgets
 * For partner reservation website
 * OpenAPI spec version: v0.0.1
 */
import type { CarManufacturer } from "./carManufacturer";
import type { ModelVariant } from "./modelVariant";
import type { SippTransmission } from "./sippTransmission";
import type { ShortVehicleCategory } from "./shortVehicleCategory";

export interface VehiclePublic {
  uuid?: string;
  manufacturer: CarManufacturer;
  model: string;
  modelRange?: string;
  modelVariant?: ModelVariant;
  doorsNumber?: number;
  seatsNumber?: number;
  modelYear?: number;
  transmission?: SippTransmission;
  category?: ShortVehicleCategory;
  pictureUrl?: string;
  thumbnailUrl?: string;
  adminDisplayName?: string;
  shortDisplayName?: string;
  airConditioning?: boolean;
}